WGL-5: Rendering Optimization (#9)
* WGL-5: Rendering Optimization

- Fix for performance issue related to render-target readback and uploading for each view.

- Many changes to Leia render controller, major changes are:
  - Deleted setupCanvas() and setupTextures() with logic moved to initialize()
  - initialize() has 3 extra parameters: nativeMode, maxAttribs, and enableViewSharpening.
  - Single canvas instead of two
  - SDK and app share the same WebGL context (requires care)
  - Added bindRenderTarget()

Native mode means the SDK creates render-targets and the app calls bindRenderTarget() to select each view. Non-native mode means the app creates a render-target for each view and the pixels are read-back before being sent to the API via saveTexture().

Native mode is fast. Non-native is slow.

In native mode, the SDK and app share a WebGL context and therefore the SDK has to be careful to restore any modified state during calls. The maxAttribs parameter specifies the maximum number of attributes the app uses, and is defaulted to 16.

The existing Three.js example continues to use (slow) non-native mode.

The new example3 uses (fast) native mode and is pure WebGL.

I focused on pure WebGL performance due to the rush and the need to support Vadym's viewer.

This changelist doesn't include any attempt to resolve scaling or fullscreen issues and retains existing behavior in that respect.

Warning: The sharpening shader seems wrong to me. Note that example1 uses a texture that washes-out the ill-effects of the shader. The blending is definitely wrong and I suspect that may have to do with the reported gamma issues (assuming the reported issues have the same shader logic that WebGL uses).

* Added fix for WGL-3 baseline scaling issue to local SDK

* Removed commented-out shader uniform values.

Co-authored-by: Cyrus Rohani <cyrus.rohani.nl@leiainc.com>
Co-authored-by: JoeLang_LeiaInc <joseph.lang@leiainc.com>